Fix tests

+++ Source/Provider Libraries/Subscription Handling/SubscriptionHandlingTests/SubscriptionHandler/Callbacks/MultiSubscriber.cs
@@ -26,13 +26,13 @@ namespace Tests.SubscriptionHandler.Callbacks
         {
             public IEnumerator GetEnumerator()
             {
-                yield return new TestCaseData(SubReqs.Button1, 100).SetName("Subscriber 1 to Button 1 should receive the callback");
-                yield return new TestCaseData(SubReqs.Button1Subscriber2, 100).SetName("Subscriber 2 to Button 1 should receive the callback");
+                yield return new TestCaseData(SubReqs.Button1, (short)100).SetName("Subscriber 1 to Button 1 should receive the callback");
+                yield return new TestCaseData(SubReqs.Button1Subscriber2, (short)100).SetName("Subscriber 2 to Button 1 should receive the callback");
             }
         }
 
         [TestCaseSource(typeof(TestData))]
-        public void DoTests(InputSubReq sr, int value)
+        public void DoTests(InputSubReq sr, short value)
         {
             Assert.That(_subHelper.CallbackResults[sr.Name].Value, Is.EqualTo(value), "Value should be correct");
             Assert.That(_subHelper.CallbackResults[sr.Name].BindingDescriptor, Is.EqualTo(sr.BindingDescriptor), "BindingDescriptor should match binding");
